+++ .github/workflows/go-tests.yml
@@ -23,4 +23,4 @@ jobs:
       - name: Checkout code
         uses: actions/checkout@v2
       - name: Test
-        run: go test -v ./...
+        run: RUN_DOCKER_TESTS=true go test -v ./...
