+++ internal/dbsettings/database_type.go
@@ -1,4 +1,4 @@
-package database
+package dbsettings
 
 type Type string
