+++ internal/dbsettings/runsettings/run_settings.go
@@ -1,8 +1,8 @@
 package runsettings
 
-import "clickhouse-playground/internal/database"
+import "clickhouse-playground/internal/dbsettings"
 
 // RunSettings interface define custom settings for different databases
 type RunSettings interface {
-	Type() database.Type
+	Type() dbsettings.Type
 }
