+++ internal/qrunner/dockerengine/docker_image_name.go
@@ -1,4 +1,4 @@
-package qrunner
+package dockerengine
 
 import (
 	"fmt"
