+++ internal/qrunner/dockerengine/docker_image_name_test.go
@@ -1,4 +1,4 @@
-package qrunner
+package dockerengine
 
 import (
 	"testing"
