+++ pkg/chspec/clickhouse_readiness_test.go
@@ -1,4 +1,4 @@
-package qrunner
+package chspec
 
 import (
 	"testing"
