+++ pkg/chspec/semver.go
@@ -1,4 +1,4 @@
-package chsemver
+package chspec
 
 import (
 	"strconv"
