+++ pkg/chspec/semver_test.go
@@ -1,4 +1,4 @@
-package chsemver
+package chspec
 
 import (
 	"fmt"
